Move src/main/promise-exec.ts. to src/common/utils/
